% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn.R
\docType{data}
\name{espn}
\alias{espn}
\title{ESPN Win Probabilities for NCAA Football. 2016-2018 Seasons.}
\format{
A tibble with 471867 rows and 25 variables:
}
\source{
\url{https://www.espn.com}
}
\usage{
espn
}
\description{
A dataset containing the win probabilities for every game covered by ESPN in the 2016 and 2017 seasons.
The data also contains information about the game itself, such as the teams playing, the time left in the game,
the final score, etc...
}
\keyword{datasets}
