% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecap.R
\name{plot.ecap}
\alias{plot.ecap}
\title{Plotting ECAP Object}
\usage{
\method{plot}{ecap}(x)
}
\description{
Plots diagnostic information of an ECAP object. Two plots are produced. The first plot displays the estimate of the
function g that the ecap procedure produced. The second compares the unadjusted probabiltiy estimates to the ECAP adjusted
probability estimates that were used to train the model.
}
\references{
http://faculty.marshall.usc.edu/gareth-james/Research/Probs.pdf
}
\author{
Bradley Rava. PhD Candidate at the University of Southern California's Marshall School of Business.
Department of Data Sciences and Operations.
}
