% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecap.R
\name{predict.ecap}
\alias{predict.ecap}
\title{Implimenting the ECAP procedure}
\usage{
\method{predict}{ecap}(x, new_unadjusted)
}
\value{
A vector of ECAP adjusted probability estimates.
}
\description{
Takes in an ECAP object and a new set of probability estimates that the user wishes to adjust. The model uses the
calibration from the ecap object to ECAP adjust the new probability estimates given to the function predict.
}
\references{
http://faculty.marshall.usc.edu/gareth-james/Research/Probs.pdf
}
\author{
Bradley Rava. PhD Candidate at the University of Southern California's Marshall School of Business.
Department of Data Sciences and Operations.
}
