% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecap.R
\name{print.ecap}
\alias{print.ecap}
\title{Printing ECAP Object}
\usage{
\method{print}{ecap}(x, digits, ...)
}
\arguments{
\item{x}{An object of class ecap.}

\item{digits}{The number of significant digits that should be displayed.}

\item{...}{Additional arguments}
}
\description{
Prints summary information about the ECAP object
}
\references{
http://faculty.marshall.usc.edu/gareth-james/Research/Probs.pdf
}
\author{
Bradley Rava, Peter Radchenko and Gareth M. James.
}
