% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecap.R
\name{print.ecap}
\alias{print.ecap}
\title{Printing ECAP Object}
\usage{
\method{print}{ecap}(x, digits = 4)
}
\description{
Prints summary information about the ECAP object
}
\references{
http://faculty.marshall.usc.edu/gareth-james/Research/Probs.pdf
}
\author{
Bradley Rava. PhD Candidate at the University of Southern California's Marshall School of Business.
Department of Data Sciences and Operations.
}
