% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecap.R
\name{summary.ecap}
\alias{summary.ecap}
\title{Summary of ECAP Object}
\usage{
\method{summary}{ecap}(x, digits = 4)
}
\arguments{
\item{x}{An object of class ecap.}

\item{digits}{The number of significant digits that should be displayed.}
}
\description{
Prints summary information about the calibration of an ECAP object.
*** Denotes that one of the parameter estimates has hit the end of the given grid of tuning parameters.
The grid can be adjusted in the ecap function.
}
\references{
http://faculty.marshall.usc.edu/gareth-james/Research/Probs.pdf
}
\author{
Bradley Rava, Peter Radchenko and Gareth M. James.
}
